just a try, merge remote from beta git (reverted from commit 9ae0c4046e74409346b18df0a060788ac906ca45)

+++ lib/actions/action.h
@@ -54,7 +54,7 @@ private:
 		int m_prev_seen_make_key;
 	};
 
-	std::multimap<long long int, eActionBinding> m_bindings;
+	std::multimap<int, eActionBinding> m_bindings;
 
 	friend struct compare_string_keybind_native;
 	struct eNativeKeyBinding
